Add a little anonymous stats tracker just so I can see if people are using it. It curls a php script that (for now) mails me input like this:
----
Somebody used the happiness shell helper script.

Command: check_import
----

If a second arg is passed to the anon tracker function, that's appended to the email as an error message. The only other data available is what gets logged in the web server logs (IP, user agent, time, etc.).

Think I may turn on a first-run "can I track stats?" prompt too.

This commit also adds a check to see if we've executed a valid command, and the script prints a usage message if not.